+++ back-end/src/main/java/com/tdd/backend/auth/data/JwtTokenPairResponse.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.auth;
+package com.tdd.backend.auth.data;
 
 import lombok.Builder;
 import lombok.Getter;
