+++ back-end/src/main/java/com/tdd/backend/auth/exception/ExpiredATKException.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.auth;
+package com.tdd.backend.auth.exception;
 
 public class ExpiredATKException extends RuntimeException {
 	private static final String MESSAGE = "액세스 토큰이 만료되었습니다. 재발급해주세요.";
