+++ back-end/src/main/java/com/tdd/backend/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.auth;
+package com.tdd.backend.auth.exception;
 
 public class InvalidTokenException extends RuntimeException {
