+++ back-end/src/main/java/com/tdd/backend/user/data/User.java
@@ -1,11 +1,10 @@
-package com.tdd.backend.user;
+package com.tdd.backend.user.data;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
 import com.tdd.backend.post.model.Appointment;
-import com.tdd.backend.user.data.UserCreate;
 
 import lombok.Builder;
 import lombok.Getter;
