+++ back-end/src/main/java/com/tdd/backend/user/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.user;
+package com.tdd.backend.user.repository;
 
 import java.util.Optional;
 
@@ -6,6 +6,8 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import com.tdd.backend.user.data.User;
 
 @Repository
 public interface UserRepository extends CrudRepository<User, Long> {
